Refactors deploy code to be able to run arbitrary commands on the cluster
